internet-monitoring/pipeline/requirements.txt
pyyaml>=6.0
websockets>=11.0,<13.0
